greenlakecity/kafka_consumer_check/requirements.txt
flask
kafka-python
